requirements.txt
requests
tencentcloud-sdk-python==3.0.1131